backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.12.0
numpy==2.2.5
scikit-learn==1.6.1
pypdf==4.0.1
python-docx==1.1.0
tiktoken==0.9.0
